chore(genesis): cleanup system config
