+++ crates/genesis/Cargo.toml
@@ -23,6 +23,7 @@ alloy-eips.workspace = true
 
 # Misc
 thiserror.workspace = true
+derive_more.workspace = true
 
 # `revm` feature
 revm = { workspace = true, optional = true, features = ["optimism"] }
